Refactor purchase order summary updates and enhance quotation item price calculation

+++ actions/po-list/delete/index.ts
@@ -4,21 +4,38 @@ import { createSafeAction } from "@/lib/create-safe-action";
 import { InputType, ReturnType } from "./types";
 import { schema } from "./schema";
 import { revalidatePath } from "next/cache";
+import { updatePurchaseOrderSummary } from "@/app/services/service.purchase-order";
 
 const handler = async (data: InputType): Promise<ReturnType> => {
   try {
+    // find the purchase order id
+    const purchaseOrderItemId = data.id;
+    const purchaseOrderItem = await db.purchaseOrderItem.findUnique({
+      where: {
+        id: purchaseOrderItemId,
+      },
+    });
+
+    if (!purchaseOrderItem) {
+      return {
+        error: "Purchase order item not found.",
+      };
+    }
 
     await db.item.deleteMany({
       where: {
-        purchaseOrderItemId: data.id,
+        purchaseOrderItemId,
       },
     });
 
     await db.purchaseOrderItem.delete({
       where: {
-        id: data.id,
+        id: purchaseOrderItemId,
       },
     });
+
+    // update purchase order summary
+    await updatePurchaseOrderSummary(purchaseOrderItem.purchaseOrderId);
 
     revalidatePath("/purchase-orders/[id]");
